+++ src/ampio_mqtt/__init__.py
@@ -25,4 +25,4 @@ __all__ = [
     "module_model",
 ]
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"